CUMULUS-337 Move createQueue to common/aws to fix circular dependency
