+++ tests/sftp_pdr_parse_ingest.js
@@ -1,6 +1,5 @@
 'use strict';
 
-const path = require('path');
 const test = require('ava');
 const fs = require('fs-extra');
 const {
@@ -10,8 +9,9 @@ const {
   deleteCMAFromTasks,
   messageBuilder
 } = require('../packages/integration-tests/local');
-const { randomString, createQueue } = require('../packages/common/test-utils');
+const { randomString } = require('../packages/common/test-utils');
 const {
+  createQueue,
   recursivelyDeleteS3Bucket,
   s3,
   sqs,
